fix: change REAL to NUMERIC(8, 2)

+++ app/db/scripts/tables/service.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS service (
     id SERIAL PRIMARY KEY,
     description TEXT NOT NULL,
-    cost REAL NOT NULL
+    cost NUMERIC(8, 2) NOT NULL
 );
 
 CREATE INDEX idx_description ON service (description);
